# Work_Assignment/AppServer/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "PlugGoBeyond",
    "description" : "Rest API of the mobile app PlugGoBeyond of the project course LI IV in the program Mestrado Integrado em Engenharia Informática",
    "termsOfService" : "http://swagger.io/terms",
    "contact" : {
      "email" : "a34900/a82888/a85731/a86618/a89982@alunos.uminho.pt"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
...
  "servers" : [ {
    "url" : "/"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/chargingStations" : {
      "get" : {
        "tags" : [ "chargingStations" ],
        "summary" : "List all charging stations.",
        "operationId" : "findChargingStation",
        "responses" : {
          "200" : {
            "description" : "Success.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ChargingStation"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Error during operation."
          }
        }
      },
      "post" : {
        "tags" : [ "chargingStations" ],
        "summary" : "Add a new charging station into the system.",
        "operationId" : "addChargingStation",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChargingStation"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Success."
          },
          "400" : {
            "description" : "Invalid parameter."
          }
        }
      }
    },
    "/chargingStations/filterByIds" : {
      "get" : {
        "tags" : [ "chargingStations" ],
        "summary" : "Find all charging stations by the ids given.",
        "description" : "Multiple charging stations can be requested to be given. It's only required that they have a comma token to separate them.",
        "operationId" : "findChargingStationsByIds",
        "parameters" : [ {
          "name" : "ids",
          "in" : "query",
          "description" : "Ids of the charging stations.",
          "required" : true,
          "style" : "form",
          "explode" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ChargingStation"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid ids."
          }
        }
      }
    },
    "/chargingStations/{chargingStationId}" : {
      "get" : {
        "tags" : [ "chargingStations" ],
        "summary" : "Find a charging station by a given id.",
        "description" : "Returns the charging station that matches the id.",
        "operationId" : "getStationById",
        "parameters" : [ {
          "name" : "chargingStationId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ChargingStation"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid id."
          },
          "404" : {
            "description" : "Charging station not found."
          }
        }
      },
      "put" : {
        "tags" : [ "chargingStations" ],
        "summary" : "Updates a charging station.",
        "operationId" : "updateStation",
        "parameters" : [ {
          "name" : "chargingStationId",
          "in" : "path",
          "description" : "Charging station id that needs to be updated.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChargingStation"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "400" : {
            "description" : "Invalid id."
          },
          "404" : {
            "description" : "Charging station not found."
          }
        }
      },
      "delete" : {
        "tags" : [ "chargingStations" ],
        "summary" : "Removes a charging station.",
        "operationId" : "deleteChargingStation",
        "parameters" : [ {
          "name" : "chargingStationId",
          "in" : "path",
          "description" : "Charging station id that needs to be removed.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Invalid id"
          },
          "404" : {
            "description" : "Charging station not found."
          }
        }
      }
    },
    "/chargingStations/findByStatus" : {
      "get" : {
        "tags" : [ "chargingStations" ],
        "summary" : "Find all charging stations by a given status.",
        "description" : "Multiple status can be requested to be filted. It's only required that they have a comma token to separate them.",
        "operationId" : "findChargingStationByStatus",
        "parameters" : [ {
          "name" : "status",
          "in" : "query",
          "description" : "Status to filter the charging stations.",
          "required" : true,
          "style" : "form",
          "explode" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "UNKNOWN", "AVAILABLE", "NOT_AVAILABLE", "IN_USE", "UNDER_CONSTRUCTION" ]
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ChargingStation"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid status."
          }
        }
      }
    },
    "/chargingStations/findByFastConnector" : {
      "get" : {
        "tags" : [ "chargingStations" ],
        "summary" : "Find all charging stations with a fast connector.",
        "operationId" : "findFastChargingStations",
        "responses" : {
          "200" : {
            "description" : "Success.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ChargingStation"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Error during operation"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Connector" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string",
            "description" : "Status of the connector",
            "enum" : [ "UNKNOWN", "AVAILABLE", "NOT_AVAILABLE", "IN_USE", "UNDER_CONSTRUCTION" ]
          },
          "powerKw" : {
            "type" : "number",
            "format" : "double"
          },
          "amps" : {
            "type" : "number",
            "format" : "int32"
          },
          "voltage" : {
            "type" : "number",
            "format" : "int32"
          },
          "phase" : {
            "type" : "string",
            "description" : "Phase type of the connector",
            "enum" : [ "UNKNOWN", "SINGLE_PHASE_AC", "THREE_PHASE_AC", "DC" ]
          },
          "connectorType" : {
            "type" : "string",
            "description" : "Type of connector",
            "enum" : [ "UNKNOWN", "CHADEMO", "CCS", "TYPE2", "TYPE2_TETHERED", "TESLA_DESTINATION_CHARGER", "TESLA_SUPERCHARGER" ]
          },
          "rate" : {
            "type" : "string",
            "description" : "Charging rate of the connector",
            "enum" : [ "NORMAL", "FAST" ]
          }
        },
        "example" : {
          "phase" : "UNKNOWN",
          "powerKw" : 5.637376656633329,
          "amps" : 2.3021358869347655,
          "connectorType" : "UNKNOWN",
          "rate" : "NORMAL",
          "id" : "id",
          "status" : "UNKNOWN",
          "voltage" : 7.061401241503109
        }
      },
      "Position" : {
        "type" : "object",
        "properties" : {
          "latitude" : {
            "type" : "number",
            "format" : "double"
          },
          "longitude" : {
            "type" : "number",
            "format" : "double"
          }
        },
        "example" : {
          "latitude" : 6.027456183070403,
          "longitude" : 1.4658129805029452
        }
      },
      "ChargingStationOperator" : {
        "type" : "object",
        "properties" : {
          "operatorName" : {
            "type" : "string"
          },
          "websiteURL" : {
            "type" : "string",
            "format" : "uri"
          }
        },
        "example" : {
          "websiteURL" : "http://example.com/aeiou",
          "operatorName" : "operatorName"
        }
      },
      "ChargingStation" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "location" : {
            "$ref" : "#/components/schemas/Position"
          },
          "chargingStationOperator" : {
            "$ref" : "#/components/schemas/ChargingStationOperator"
          },
          "status" : {
            "type" : "string",
            "description" : "Status of the charging station",
            "enum" : [ "UNKNOWN", "AVAILABLE", "NOT_AVAILABLE", "IN_USE", "UNDER_CONSTRUCTION" ]
          },
          "priceByActivation" : {
            "type" : "number",
            "format" : "double"
          },
          "priceByMinute" : {
            "type" : "number",
            "format" : "double"
          },
          "priceByKwh" : {
            "type" : "number",
            "format" : "double"
          },
          "waitingToCharge" : {
            "type" : "number",
            "format" : "int32"
          },
          "chargingSockets" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Connector"
            }
          }
        },
        "example" : {
          "chargingSockets" : [ {
            "phase" : "UNKNOWN",
            "powerKw" : 5.637376656633329,
            "amps" : 2.3021358869347655,
            "connectorType" : "UNKNOWN",
            "rate" : "NORMAL",
            "id" : "id",
            "status" : "UNKNOWN",
            "voltage" : 7.061401241503109
          }, {
            "phase" : "UNKNOWN",
            "powerKw" : 5.637376656633329,
            "amps" : 2.3021358869347655,
            "connectorType" : "UNKNOWN",
            "rate" : "NORMAL",
            "id" : "id",
            "status" : "UNKNOWN",
            "voltage" : 7.061401241503109
          } ],
          "chargingStationOperator" : {
            "websiteURL" : "http://example.com/aeiou",
            "operatorName" : "operatorName"
          },
          "priceByMinute" : 6.027456183070403,
          "waitingToCharge" : 5.962133916683182,
          "location" : {
            "latitude" : 6.027456183070403,
            "longitude" : 1.4658129805029452
          },
          "priceByActivation" : 0.8008281904610115,
          "id" : "id",
          "priceByKwh" : 1.4658129805029452,
          "status" : "UNKNOWN"
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}